requirements.txt
Babel==2.11.0
python-dotenv==0.21.0
requests==2.26.0
python_telegram_bot==13.13
telegram==0.0.1